fix all

+++ EPMS.Domain/Interface/IService/ILevelService.cs
@@ -6,8 +6,8 @@ namespace EPMS.Domain.Interfaces;
 public interface ILevelService
 {
     Task<SuccessResponse<IEnumerable<LevelDto>>> GetAllAsync();
-    Task<SuccessResponse<LevelDto>> GetByIdAsync(int id);
-    Task<SuccessResponse<int>> CreateAsync(CreateLevelDto dto);
-    Task<SuccessResponse> UpdateAsync(int id, UpdateLevelDto dto);
-    Task<SuccessResponse> DeleteAsync(int id);
+    Task<SuccessResponse<LevelDto>> GetByIdAsync(long id);
+    Task<SuccessResponse<long>> CreateAsync(CreateLevelDto dto);
+    Task<SuccessResponse> UpdateAsync(long id, UpdateLevelDto dto);
+    Task<SuccessResponse> DeleteAsync(long id);
 }
