+++ EPMS.Shared/DTOs/CategoryDTOs/CreateCategoryDto.cs
@@ -6,5 +6,5 @@ public record CreateCategoryDto
     public string Code { get; init; } = string.Empty;
     public string Name { get; init; } = string.Empty;
     public string? Description { get; init; }
-    public int? ParentId { get; init; }
+    public long? ParentId { get; init; }
 }
